Remove chain name from logic

+++ core/src/main/kotlin/co/nilin/opex/chainscan/core/model/DepositResult.kt
@@ -1,6 +1,8 @@
 package co.nilin.opex.chainscan.core.model
 
+import java.math.BigInteger
+
 data class DepositResult(
-    val latestBlock: Long,
+    val latestBlock: BigInteger,
     val deposits: List<Deposit>
 )
