+++ core/src/main/kotlin/co/nilin/opex/chainscan/core/spi/ChainEndpointProxy.kt
@@ -1,9 +1,10 @@
 package co.nilin.opex.chainscan.core.spi
 
 import co.nilin.opex.chainscan.core.model.ChainSyncRecord
+import java.math.BigInteger
 
 interface ChainEndpointProxy {
-    data class DepositFilter(val startBlock: Long?, val endBlock: Long?, val tokenAddresses: List<String>?)
+    data class DepositFilter(val startBlock: BigInteger?, val endBlock: BigInteger?, val tokenAddresses: List<String>?)
 
     suspend fun syncTransfers(filter: DepositFilter): ChainSyncRecord
 }
